pyproject.toml
[build-system]
build-backend = "hatchling.build"

requires = [ "hatchling>=1.26" ]

[project]
name = "aiohttp-underscore-apis"
version = "2025.10.24a1"
description = "Inspect and manage aiohttp web application like Elasticsearch"
readme = "README.md"
license = "MIT"
license-files = [ "LICENSE" ]

requires-python = ">=3.10"

classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
]
dependencies = [
  "aiohttp",
  "pyyaml",
  "tabulate",
  "webargs",
]
optional-dependencies.dev = [
  "types-pyyaml",
  "types-tabulate",
]

[tool.black]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79
